refactor(web): repoint Sanity importers, delete the lib/sanity barrels + Studio embed (SP2-C Task 4)
Repoint at HEAD (re-derived): 31 imports @/lib/sanity/queries ->
@/lib/content/queries; 15 imports @/lib/sanity/types -> @superteam-lms/types;
4 route imports @/lib/sanity/admin-mutations -> @/lib/content/deployment-writes
(the writers already live there since SP2-B Task 6 — admin-mutations was only
a re-export plus a dead Sanity write client whose sole caller, the group-C
gateway, died in Task 3; its test covered only that dead half).

Delete lib/sanity/{client,queries,types,admin-mutations}.ts + __tests__,
app/studio/ (NextStudio embed) + apps/web/sanity.config.ts (they import
next-sanity/@sanity/vision/@superteam-lms/sanity — unusable post-SP2-C and
required for the gate).

GATE: git grep -lE "from ['\"](@/lib/sanity|@sanity/|next-sanity)" apps/web/src
-> ZERO. next build green. Suite 59 files / 499 tests pass; tsc + lint clean.
