+++ apps/web/src/app/[locale]/(platform)/courses/[slug]/lessons/[id]/blocks/quiz-block.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useCallback, useEffect, useState } from "react";
-import type { QuizBlockData } from "@/lib/sanity/types";
+import type { QuizBlockData } from "@superteam-lms/types";
 import type { BlockRenderProps } from "./types";
 
 /**
