+++ apps/web/src/app/[locale]/(platform)/courses/page.tsx
@@ -1,5 +1,5 @@
 import { CourseCatalogClient } from "./courses-client";
-import { getAllCourses, getAllLearningPaths } from "@/lib/sanity/queries";
+import { getAllCourses, getAllLearningPaths } from "@/lib/content/queries";
 
 export default async function CoursesPage() {
   const [courses, learningPaths] = await Promise.all([
