+++ apps/web/src/app/api/lessons/complete/__tests__/gate.test.ts
@@ -35,7 +35,7 @@ vi.mock("@/lib/supabase/admin", () => ({
   }),
 }));
 
-vi.mock("@/lib/sanity/queries", () => ({
+vi.mock("@/lib/content/queries", () => ({
   getLessonByIdForGrading,
   getCourseById,
 }));
