--- apps/web/src/app/studio/[[...tool]]/page.tsx
@@ -1,10 +1,0 @@
-import { NextStudio } from "next-sanity/studio";
-import config from "../../../../sanity.config";
-
-export const dynamic = "force-static";
-
-export { metadata, viewport } from "next-sanity/studio";
-
-export default function StudioPage() {
-  return <NextStudio config={config} />;
-}
