--- apps/web/src/app/studio/layout.tsx
@@ -1,11 +1,0 @@
-export default function StudioLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <html lang="en">
-      <body style={{ margin: 0 }}>{children}</body>
-    </html>
-  );
-}
